+++ CHANGES.md
@@ -14,8 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Initialize data loaders for training and validation dataset once per fit call instead of once per epoch ([migration guide](https://skorch.readthedocs.io/en/stable/user/FAQ.html#migration-from-0-11-to-0-12))
+- It is now possible to call `np.asarray` with `SliceDataset`s (#858)
 
 ### Fixed
+- Fix a bug in `SliceDataset` that prevented it to be used with `to_numpy` (#858)
 
 ## [0.11.0] - 2021-10-11